remake the enemy script and player script and add guns and more zombies
make ascript for the enemy attack and animation and fix the old enemymaneger Fix the problem whit the player and weapon add more guns and a switching system add more zombies and place them in the first stage make the health and the damage float instade of int also add an explotion script and set the lighting for the first stage second mission

+++ Assets/Scripts/DangerZoneScript.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class DangerZoneScript : MonoBehaviour
 {
-    [SerializeField] private int _damagePerSecond;
+    [SerializeField] private float _damagePerSecond;
     [SerializeField] private GameObject _player;
     [SerializeField] private LayerMask _targetObject;
     [SerializeField] private float _radius;
